// apps/core-service/src/app/modules/chat/types/sendbird/SendbirdError.ts
//=============================================================================
//Licensed Materials - Property of Bambee
//(C) Copyright Bambee 2022
//All Rights Reserved
//=============================================================================

export interface SendbirdError {
  message: string;
  code: number;
  error: true;
}